Fix importing data to test file and repass test
