+++ test/testIngredientData.js
@@ -1,0 +1,173 @@
+const ingredientsData = [
+  {
+    "id": 20081,
+    "name": "wheat flour",
+    "estimatedCostInCents": 142
+  },
+  {
+    "id": 18372,
+    "name": "bicarbonate of soda",
+    "estimatedCostInCents": 582
+  },
+  {
+    "id": 1123,
+    "name": "eggs",
+    "estimatedCostInCents": 472
+  },
+  {
+    "id": 19335,
+    "name": "sucrose",
+    "estimatedCostInCents": 902
+  },
+  {
+    "id": 19206,
+    "name": "instant vanilla pudding",
+    "estimatedCostInCents": 660
+  },
+  {
+    "id": 19334,
+    "name": "brown sugar",
+    "estimatedCostInCents": 559
+  },
+  {
+    "id": 2047,
+    "name": "salt",
+    "estimatedCostInCents": 280
+  },
+  {
+    "id": 1012047,
+    "name": "fine sea salt",
+    "estimatedCostInCents": 528
+  },
+  {
+    "id": 10019903,
+    "name": "semi sweet chips",
+    "estimatedCostInCents": 253
+  },
+  {
+    "id": 1145,
+    "name": "unsalted butter",
+    "estimatedCostInCents": 617
+  },
+  {
+    "id": 2050,
+    "name": "vanilla",
+    "estimatedCostInCents": 926
+  },
+  {
+    "id": 1009016,
+    "name": "apple cider",
+    "estimatedCostInCents": 468
+  },
+  {
+    "id": 9003,
+    "name": "apple",
+    "estimatedCostInCents": 207
+  },
+  {
+    "id": 20027,
+    "name": "corn starch",
+    "estimatedCostInCents": 236
+  },
+  {
+    "id": 1002046,
+    "name": "dijon style mustard",
+    "estimatedCostInCents": 619
+  },
+  {
+    "id": 11215,
+    "name": "whole garlic clove",
+    "estimatedCostInCents": 220
+  },
+  {
+    "id": 1012046,
+    "name": "whole grain dijon mustard",
+    "estimatedCostInCents": 867
+  },
+  {
+    "id": 19911,
+    "name": "maple",
+    "estimatedCostInCents": 349
+  },
+  {
+    "id": 16112,
+    "name": "miso",
+    "estimatedCostInCents": 978
+  },
+  {
+    "id": 10010062,
+    "name": "pork chop",
+    "estimatedCostInCents": 834
+  },
+  {
+    "id": 1102047,
+    "name": "s&p",
+    "estimatedCostInCents": 524
+  },
+  {
+    "id": 16124,
+    "name": "soy sauce",
+    "estimatedCostInCents": 486
+  },
+  {
+    "id": 1016168,
+    "name": "sriracha sauce",
+    "estimatedCostInCents": 576
+  },
+  {
+    "id": 1002030,
+    "name": "black pepper",
+    "estimatedCostInCents": 441
+  },
+  {
+    "id": 1001,
+    "name": "butter",
+    "estimatedCostInCents": 618
+  },
+  {
+    "id": 4582,
+    "name": "oil",
+    "estimatedCostInCents": 807
+  },
+  {
+    "id": 2031,
+    "name": "red pepper powder",
+    "estimatedCostInCents": 583
+  },
+  {
+    "id": 5100,
+    "name": "chicken wing",
+    "estimatedCostInCents": 593
+  },
+  {
+    "id": 2009,
+    "name": "red chili powder",
+    "estimatedCostInCents": 499
+  },
+  {
+    "id": 1022020,
+    "name": "garlic powder",
+    "estimatedCostInCents": 344
+  },
+  {
+    "id": 6168,
+    "name": "tabasco sauce",
+    "estimatedCostInCents": 859
+  },
+  {
+    "id": 9176,
+    "name": "mangoes",
+    "estimatedCostInCents": 425
+  },
+  {
+    "id": 2026,
+    "name": "onion powder",
+    "estimatedCostInCents": 597
+  },
+  {
+    "id": 1042047,
+    "name": "seasoned salt",
+    "estimatedCostInCents": 334
+  },]
+
+  export default ingredientsData
